requirements.txt
Django<=3.1.99
djangorestframework<=3.11.99
geopy<=2.2.99